isolate Searchbar and UserControlMenu as stand-alone components
