+++ src/components/components.js
@@ -7,9 +7,11 @@ import TrashBin from './TrashBin';
 import NavBar from '../containers/NavBar';
 import GlobalAuthState from '../scripts/Auth/GlobalAuthState';
 import SignInWithGoogleBTN from './SignInWithGoogleBTN';
+import SearchBar from './SearchBar';
+import UserControlMenu from './UserControlMenu';
 import Footer from './Footer';
 import MySpinner from './MySpinner';
 
 export {
-  File, UploadArea, TrashBin, NavBar, GlobalAuthState, UploadButton, SignInWithGoogleBTN, Footer, MySpinner,
+  File, UploadArea, TrashBin, NavBar, GlobalAuthState, UploadButton, SignInWithGoogleBTN, SearchBar, UserControlMenu, Footer, MySpinner,
 };
